Updating the boost sources to the latest version and also adding new libraries to BoostParts; also removed some unused files

+++ src/BoostParts/boost/smart_ptr/detail/yield_k.hpp
@@ -55,7 +55,7 @@ namespace detail
 {
 
 #if !defined( BOOST_USE_WINDOWS_H )
-  extern "C" void __stdcall Sleep( unsigned ms );
+  extern "C" void __stdcall Sleep( unsigned long ms );
 #endif
 
 inline void yield( unsigned k )
